some more corrections and rename of py_models.py
